+++ .gitignore
@@ -1,4 +1,5 @@
 yarn.lock
+.DS_Store
 # Logs
 logs
 *.log
