+++ README.md
@@ -15,5 +15,5 @@ docker-compose up
 Para testar a aplicação rode o comando abaixo:
 
 ```
-docker-compose run --rm -e NODE_ENV=test sensors_data_api bash -c  "yarn && yarn jest --coverage --forceExit"
+docker-compose run --rm -e NODE_ENV=test sensors_data_api bash -c  "yarn && yarn jest --coverage --forceExit --runInBand"
 ```
