+++ src/app.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const routes = require('./routes');
+const cors = require('cors')
+
+const app = express();
+app.use(express.json());
+app.use(cors());
+app.use(routes);
+
+
+module.exports = app;
